move UserProfile components into individual files
