+++ frontend/src/app/components/UserProfile.module.css
@@ -42,355 +42,3 @@
   gap: var(--Spacing-md, 24px);
   align-self: stretch;
 }
-
-/* PROFILE HEADER STYLES */
-
-.profilePicture {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: var(--Background-3, #e9ebed);
-  color: var(--Text-3, #3a3c3d);
-}
-
-.profilePicture.large {
-  font-size: 48px;
-  min-width: 96px;
-  height: 96px;
-  border-radius: 48px;
-}
-.profilePicture.small {
-  width: 28px;
-  height: 28px;
-  min-width: 28px;
-  border-radius: 14px;
-}
-
-.profileHeader {
-  display: flex;
-  width: 1024px;
-  justify-content: space-between;
-}
-
-.profileHeader .profileContent {
-  display: flex;
-  gap: 24px;
-}
-
-.profileHeader .userInfo {
-  display: flex;
-  flex-direction: column;
-  gap: var(--Spacing-s, 8px);
-  justify-content: center;
-}
-
-.profileHeader .userInfoHeader {
-  display: flex;
-  align-items: center;
-  gap: var(--Spacing-sm, 16px);
-}
-
-.profileHeader .userFullName {
-  color: var(--Text-3, #3a3c3d);
-
-  /* title medium */
-  font-family: "Albert Sans";
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-.profileHeader .userMetadata {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--Spacing-s, 8px);
-  flex-wrap: wrap;
-}
-
-.profileHeader .userMetadata .metadataSubsection {
-  display: flex;
-}
-
-.profileHeader .userMetadata .metadataSubsection * {
-  width: fit-content;
-  color: var(--Text-1, #60696f);
-  /* body large */
-  font-family: "Albert Sans";
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-.profileHeader .userMetadata .smallMetadata {
-  font-size: 14px;
-}
-
-.profileHeader .profileContentControls {
-  display: flex;
-  padding: var(--Spacing-s, 8px) 0px;
-  gap: var(--Spacing-s, 8px);
-}
-
-/* USER LIST STYLES */
-
-.userList {
-  display: flex;
-  width: 100%;
-  padding: 24px 24px var(--Spacing-l, 32px) 24px;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--Spacing-sm, 16px);
-  height: fit-content;
-
-  border-radius: 8px;
-  border: 1px solid var(--Stroke-2, #d0d7de);
-  background: #fff;
-}
-
-.userList.minimized {
-  width: 256px;
-  min-width: 256px;
-}
-
-.userList .userListHeader {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.userListHeading {
-  color: var(--Text-3, #3a3c3d);
-  font-family: "Albert Sans";
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.userList .addUser {
-  display: flex;
-  height: 34px;
-  padding: var(--Spacing-sm, 16px) 12px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  width: fit-content;
-
-  border-radius: var(--Spacing-sm, 16px);
-  border: 1px solid var(--Primary, #057e6f);
-  background: var(--Primary, #057e6f);
-}
-
-.userList .addUser:hover {
-  cursor: pointer;
-}
-
-.userList .removeUser:hover {
-  cursor: pointer;
-}
-
-.userListContent {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--Spacing-md, 24px);
-  align-self: stretch;
-}
-
-.userList .programSection {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 16px;
-  align-self: stretch;
-}
-
-.userList .programSectionHeader {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  align-items: center;
-}
-.programSectionHeaderSectionInfo {
-  display: flex;
-  gap: 8px;
-  height: fit-content;
-}
-
-.userList .userInfo {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 4px;
-  align-self: stretch;
-}
-
-.userList .userInfo:not(:last-child) {
-  border-bottom: 1px solid var(--Stroke-1, #edf1f5);
-  padding-bottom: var(--Spacing-sm, 16px);
-}
-
-.userList .fullName {
-  color: var(--Text-2, #4d5358);
-  /* body medium */
-  font-family: "Albert Sans";
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-.userList .email {
-  color: var(--Text-1, #60696f);
-  /* body small */
-  font-family: "Albert Sans";
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-
-.userList .unassigned {
-  color: var(--Danger-Emphasis, #b80037);
-  /* body medium */
-  font-family: "Albert Sans";
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-/* VOLUNTEER NOTES STYLES */
-
-.volunteerNotes {
-  display: flex;
-  padding: 24px 24px var(--Spacing-l, 32px) 24px;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--Spacing-sm, 16px);
-  align-self: stretch;
-  width: 100%;
-  height: fit-content;
-
-  border-radius: 8px;
-  border: 1px solid var(--Stroke-2, #d0d7de);
-  background: #fff;
-}
-.volunteerNotes .noteHeader {
-  display: flex;
-  height: 24px;
-  align-items: center;
-  color: var(--Text-3, #3a3c3d);
-  font-family: "Albert Sans";
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-.volunteerNotes .noteSection {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--Spacing-md, 24px);
-  align-self: stretch;
-}
-.volunteerNotes .postNoteSection {
-  display: flex;
-  align-items: flex-start;
-  gap: 18px;
-  align-self: stretch;
-}
-.volunteerNotes .postNoteInputField {
-  display: flex;
-  padding: 0px var(--Spacing-sm, 16px);
-  justify-content: space-between;
-  align-items: center;
-  flex: 1 0 0;
-  align-self: stretch;
-
-  border-radius: 4px;
-  border: 1px solid var(--Stroke-2, #d0d7de);
-}
-
-.volunteerNotes .postNoteButton {
-  all: unset;
-  display: flex;
-  height: 39px;
-  padding: 8px 24px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-
-  border-radius: 8px;
-  background: var(--Primary, #057e6f);
-
-  color: var(--Neutral-Base-White, #fafafa);
-  font-family: "Albert Sans";
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-
-  height: fit-content;
-}
-.volunteerNotes .postNoteButton:hover {
-  cursor: pointer;
-}
-
-.volunteerNotes .postedNotes {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--Spacing-sm, 16px);
-}
-.volunteerNotes .postedNote {
-  display: flex;
-  align-items: flex-start;
-  gap: 8px;
-}
-.volunteerNotes .noteContent {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 8px;
-}
-.volunteerNotes .noteHeader {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-.volunteerNotes .noteAuthor {
-  color: var(--Text-3, #3a3c3d);
-
-  /* body medium */
-  font-family: "Albert Sans";
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-.volunteerNotes .notePostedDate {
-  color: var(--Text-1, #60696f);
-
-  /* caption */
-  font-family: var(--font-inconsolata);
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-.volunteerNotes .noteBody {
-  color: var(--Text-3, #3a3c3d);
-
-  /* body small */
-  font-family: "Albert Sans";
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
