+++ frontend/src/app/components/UserProfile.tsx
@@ -1,25 +1,18 @@
 "use client";
-import Image from "next/image";
-import { usePathname, useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
 import {
-  AssignedProgram as AssignedProgramEnum,
-  ProfileComment,
-  ProfileCommentPostRequest,
   Role as RoleEnum,
   UserProfile as UserProfileType,
-  getComments,
   getUserProfile,
-  postComment,
 } from "../api/profileApi";
 
 import NavigateBack from "./NavigateBack";
 import styles from "./UserProfile.module.css";
 
-import { Program } from "@/app/components//Program";
-import { Button } from "@/app/components/Button";
-import { Role } from "@/app/components/Role";
+import { ProfileHeader } from "@/app/components/ProfileHeader";
+import { VolunteerNotes } from "@/app/components/VolunteerNotes";
+import { UserList } from "@/app/components/UserList";
 
 interface ProfileRenderingContext {
   invalidContext: boolean;
@@ -193,295 +186,4 @@ export default function UserProfile({ userId }: { userId: string }) {
         )}
       </>
     );
-}
-
-function ProfilePicture(params: { firstName?: string; size?: string }) {
-  const { firstName, size } = params;
-  const sizeClass = size === "small" ? "small" : "large";
-
-  const firstLetter = firstName && firstName.length > 0 ? firstName[0] : "?";
-  return <div className={`${styles.profilePicture} ${styles[sizeClass]}`}>{firstLetter}</div>;
-}
-
-function Divider() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20" fill="none">
-      <path
-        d="M9.75 1C9.94891 1 10.1397 1.07902 10.2803 1.21967C10.421 1.36032 10.5 1.55109 10.5 1.75V18.25C10.5 18.4489 10.421 18.6397 10.2803 18.7803C10.1397 18.921 9.94891 19 9.75 19C9.55109 19 9.36032 18.921 9.21967 18.7803C9.07902 18.6397 9 18.4489 9 18.25V1.75C9 1.55109 9.07902 1.36032 9.21967 1.21967C9.36032 1.07902 9.55109 1 9.75 1Z"
-        fill="#60696F"
-      />
-    </svg>
-  );
-}
-
-function ProfileHeader(params: {
-  firstName: string | undefined;
-  lastName: string | undefined;
-  role: RoleEnum | undefined;
-  assignedPrograms: AssignedProgramEnum[] | undefined;
-  yearJoined?: number | undefined;
-  age?: number | undefined;
-  phoneNumber?: string | undefined;
-  gender?: string | undefined;
-  email: string | undefined;
-  isProgramAndRoleEditable: boolean;
-  isPersonalProfile: boolean;
-}) {
-  const {
-    firstName,
-    lastName,
-    role,
-    assignedPrograms,
-    yearJoined,
-    age,
-    phoneNumber,
-    gender,
-    email,
-    isProgramAndRoleEditable,
-    isPersonalProfile,
-  } = params;
-  const fullName = `${firstName ?? "Unknown"} ${lastName ?? "Unknown"}`;
-  const joinedText = `Joined: ${yearJoined ?? 0}`;
-  const ageText = `Age: ${age ?? 0}`;
-  const genderText = `Gender: ${gender ?? "Unknown"}`;
-  assignedPrograms?.sort();
-  const router = useRouter();
-  const pathname = usePathname();
-  return (
-    <div className={styles.profileHeader}>
-      <div className={styles.profileContent}>
-        <ProfilePicture firstName={firstName} />
-        <div className={styles.userInfo}>
-          <div className={styles.userInfoHeader}>
-            <div className={styles.userFullName}>{fullName}</div>
-            <Role role={role} />
-            {assignedPrograms?.map((program) => <Program program={program} key={program} />)}
-          </div>
-          <div className={styles.userMetadata}>
-            <div className={styles.metadataSubsection}>
-              <div>{joinedText}</div>
-              <Divider />
-              <div>{ageText}</div>
-              <Divider />
-              <div>{genderText}</div>
-            </div>
-
-            <div className={styles.metadataSubsection}>
-              <div className={styles.smallMetadata}>{email}</div>
-              <Divider />
-              <div className={styles.smallMetadata}>{phoneNumber}</div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className={styles.profileContentControls}>
-        {isPersonalProfile ? (
-          <Button
-            text="Edit profile"
-            onClick={() => {
-              router.push(`${pathname}/edit`);
-            }}
-          />
-        ) : isProgramAndRoleEditable ? (
-          <>
-            <Button text={"Edit Program"} /> <Button text={"Change Role"} />
-          </>
-        ) : (
-          <></>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function VolunteerNotes({ userId }: { userId: string }) {
-  const [profileNotes, setProfileNotes] = useState<ProfileComment[] | undefined>([]);
-  const [currentComment, setCurrentComment] = useState<string>("");
-  // use this to trigger a re-fetch of the notes when new note posted
-  const [profileNotesChanged, setProfileNotesChanged] = useState<boolean>(false);
-
-  useEffect(() => {
-    const fetchProfileNotes = async () => {
-      const res = await getComments(userId);
-      if (res.success) {
-        return res.data;
-      }
-    };
-    fetchProfileNotes()
-      .then((res) => {
-        setProfileNotes(res);
-      })
-      .catch((err: unknown) => {
-        console.error(err);
-      });
-  }, [profileNotesChanged]);
-
-  const postProfileNote = async (comment: ProfileCommentPostRequest) => {
-    const res = await postComment(comment);
-    if (res.success) {
-      setProfileNotesChanged(!profileNotesChanged);
-    }
-  };
-  return (
-    <div className={styles.volunteerNotes}>
-      <div className={styles.noteHeader}>Notes from Volunteers</div>
-      <div className={styles.noteSection}>
-        <div className={styles.postNoteSection}>
-          <input
-            className={styles.postNoteInputField}
-            placeholder="Add a comment..."
-            value={currentComment}
-            onChange={(event) => {
-              setCurrentComment(event.target.value);
-            }}
-          ></input>
-          <button
-            onClick={(event) => {
-              event.preventDefault();
-              const comment = {
-                profileId: userId,
-                commenterId: localStorage.getItem("viewerId") ?? "67a4255fc7beaa03529393dc",
-                comment: currentComment,
-                datePosted: new Date(),
-              };
-              postProfileNote(comment)
-                .then(() => {
-                  console.log("Comment posted");
-                  setCurrentComment("");
-                })
-                .catch((err: unknown) => {
-                  console.error(err);
-                });
-            }}
-            className={styles.postNoteButton}
-          >
-            Post
-          </button>
-        </div>
-        <div className={styles.postedNotes}>
-          {profileNotes && profileNotes.length > 0 ? (
-            profileNotes.map((comment, ind) => {
-              return (
-                <div key={ind} className={styles.postedNote}>
-                  <ProfilePicture firstName={comment.user} size="small" />
-                  <div className={styles.noteContent}>
-                    <div className={styles.noteHeader}>
-                      <div className={styles.noteAuthor}>{comment.user}</div>
-                      <div className={styles.notePostedDate}>
-                        {Math.floor(
-                          (new Date().getTime() - new Date(comment.datePosted).getTime()) /
-                            (1000 * 60 * 60 * 24),
-                        )}{" "}
-                        days ago
-                      </div>
-                    </div>
-                    <div className={styles.noteBody}>{comment.comment}</div>
-                  </div>
-                </div>
-              );
-            })
-          ) : (
-            <div>No notes</div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function UserList(params: {
-  userProfile: UserProfileType | undefined;
-  title: string;
-  editable: boolean;
-  minimized: boolean;
-}) {
-  const { title, userProfile, editable, minimized } = params;
-
-  // Users for user list
-  const emptyUserGroups: { [key: string]: UserProfileType[] } = (
-    userProfile?.assignedPrograms ?? []
-  ).reduce((accumulator: { [key: string]: UserProfileType[] }, program: string) => {
-    accumulator[program] = [];
-    return accumulator;
-  }, {});
-
-  const assignedUsers = userProfile?.assignedUsers ?? [];
-  const userGroups: { [key: string]: UserProfileType[] } = assignedUsers.reduce(
-    (accumulator, user) => {
-      (user?.assignedPrograms ?? []).forEach((program: string) => {
-        if (program in accumulator) {
-          accumulator[program].push(user);
-        }
-      });
-      return accumulator;
-    },
-    emptyUserGroups,
-  );
-
-  const sortedUserGroups: [string, UserProfileType[]][] = Object.entries(userGroups).slice().sort();
-
-  return (
-    <div className={`${styles.userList} ${minimized && styles.minimized}`}>
-      <div className={styles.userListHeader}>
-        <div className={styles.userListHeading}>{title}</div>
-        {editable && !minimized && (
-          <div className={styles.addUser}>
-            <Image
-              src="/pajamas_assignee_icon.svg"
-              width={16}
-              height={16}
-              alt="Assign User"
-            ></Image>
-          </div>
-        )}
-      </div>
-      <div className={styles.userListContent}>
-        {sortedUserGroups.map(([program, users]) => {
-          return (
-            <div key={program} className={styles.programSection}>
-              <div className={styles.programSectionHeader}>
-                <div className={styles.programSectionHeaderSectionInfo}>
-                  <Program program={program} />
-                </div>
-                {editable && minimized && (
-                  <div className={styles.addUser}>
-                    <Image
-                      src="/pajamas_assignee_icon.svg"
-                      width={16}
-                      height={16}
-                      alt="Assign User"
-                    ></Image>
-                  </div>
-                )}
-              </div>
-              {users.length > 0 ? (
-                users.map((user, ind) => {
-                  const fullName = `${user.firstName} ${user.lastName}`;
-                  return (
-                    <div key={ind} className={styles.userInfo}>
-                      <div>
-                        <div className={styles.fullName}>{fullName}</div>
-                        <div className={styles.email}>{user.email}</div>
-                      </div>
-                      {editable && (
-                        <Image
-                          src="/trash_icon.svg"
-                          width={20}
-                          height={20}
-                          alt="Remove User"
-                          className={styles.removeUser}
-                        ></Image>
-                      )}
-                    </div>
-                  );
-                })
-              ) : (
-                <div className={styles.unassigned}>Unassigned</div>
-              )}
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
 }
